Concurrency pre-merge

+++ server/src/main/java/com/tripco/t13/server/TripCalculate.java
@@ -34,7 +34,7 @@ public class TripCalculate implements Callable<int []> {
         int shortestCumulativeDistance = 0;
         trip.distances = trip.getTripDistances();
 
-        for (int distance : trip.distances) {
+        for (long distance : trip.distances) {
             shortestCumulativeDistance += distance;
         }
 
@@ -46,7 +46,7 @@ public class TripCalculate implements Callable<int []> {
 
 
         double radius = trip.options.getRadius();
-        int[][] distanceLibrary = new int[retainOriginalPlaces.size() + 1][retainOriginalPlaces.size() + 1];
+        long[][] distanceLibrary = new long[retainOriginalPlaces.size() + 1][retainOriginalPlaces.size() + 1];
         for (int i = 0; i < retainOriginalPlaces.size(); i++) {
             for (int k = 0; k < retainOriginalPlaces.size(); k++) {
                 distanceLibrary[i][k] = Distance.getDistanceNum(retainOriginalPlaces, i, k, radius);
@@ -92,7 +92,7 @@ public class TripCalculate implements Callable<int []> {
         return tempPointerPlaces;
     }
 
-    public void twoOpt(ArrayList<Location> places, int[][] distanceLibrary, int[] pPlaces) {
+    public void twoOpt(ArrayList<Location> places, long[][] distanceLibrary, int[] pPlaces) {
         boolean improvement = true;
         while (improvement) {
             improvement = false;
